LLM/requirements.txt
# 基础科学计算包
numpy>=1.21.0
scipy>=1.9.0

# 深度学习相关
cupy-cuda12x>=12.0.0  # 适配CUDA 12.8
torch>=2.1.0          # PyTorch基础包
torchvision>=0.16.0   # PyTorch视觉工具包

# 数据可视化
matplotlib>=3.5.0
seaborn>=0.12.0       # 统计数据可视化

# 数据处理
pandas>=2.0.0         # 数据分析工具
h5py>=3.8.0          # HDF5文件格式支持

# 工具包
tqdm>=4.65.0         # 进度条显示
psutil>=5.9.0        # 系统和进程监控
pillow>=10.0.0       # 图像处理

# 开发工具
ipython>=8.0.0       # 交互式Python环境
jupyter>=1.0.0       # Jupyter笔记本 